feat(home): integrate real data and use cases
* Connect `HomeViewModel` to `GetAllDestinationsUseCase` and `GetNearbyDestinationsUseCase` to replace mock data with real API calls.
* Refactor `HomeUiState` to use a more granular `UiState` wrapper for different sections (countries, recommended, nearby).
* Implement `HomeAction` and `HomeEvent` patterns for better MVI-style communication between UI and ViewModel.
* Update `HomeScreen` and its components (`RecentViewedCard`, `DestinationCard`) to support dynamic data loading and shimmer effects.
* Fix data types in `Destination` models, changing coordinates and ratings from `Int` to `Double`.

+++ core/network/src/main/java/com/example/network/dto/destinations/Destination.kt
@@ -6,9 +6,9 @@ data class Destination(
     val destinationID: Int,
     val imageUrls: List<String>,
     val interests: List<Interest>,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val name: String,
-    val rating: Int,
+    val rating: Double,
     val totalReviews: Int
 )
